requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.15.0=pypi_0
aggdraw=1.3.16=pypi_0
alabaster=0.7.13=pypi_0
art=6.0=pypi_0
astunparse=1.6.3=pypi_0
babel=2.12.1=pypi_0
beautifulsoup4=4.12.2=pypi_0
brotli=1.0.9=pypi_0
ca-certificates=2023.7.22=h56e8100_0
cached-property=1.5.2=pypi_0
cachetools=5.3.1=pypi_0
carla=0.9.13=pypi_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=pypi_0
charset-normalizer=3.2.0=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
configargparse=1.1=pypi_0
cssselect2=0.7.0=pypi_0
cudatoolkit=11.2.2=h933977f_10
cudnn=8.1.0.77=h3e0f4f4_0
cycler=0.11.0=pypi_0
docutils=0.18.1=pypi_0
flatbuffers=1.12=pypi_0
fonttools=4.38.0=pypi_0
furo=2023.3.27=pypi_0
gast=0.4.0=pypi_0
ghp-import=2.1.0=pypi_0
google-auth=2.22.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.34.1=pypi_0
h5py=3.1.0=pypi_0
html5lib=1.1=pypi_0
idna=3.4=pypi_0
imageio=2.31.1=pypi_0
imagesize=1.4.1=pypi_0
importlib-metadata=6.7.0=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.1=pypi_0
keras-nightly=2.5.0.dev2021032900=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.4=pypi_0
libsass=0.22.0=pypi_0
markdown=3.4.3=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.2.0=pypi_0
mdit-py-plugins=0.3.5=pypi_0
mdurl=0.1.2=pypi_0
mergedeep=1.3.4=pypi_0
mkdocs=1.5.2=pypi_0
myst-parser=1.0.0=pypi_0
networkx=2.6.3=pypi_0
numpy=1.19.5=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.2.0.32=pypi_0
openssl=1.1.1v=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=23.1=pypi_0
pandas=1.0.2=pypi_0
pathspec=0.11.2=pypi_0
pillow=9.5.0=pypi_0
pip=22.3.1=py37haa95532_0
platformdirs=3.10.0=pypi_0
pockets=0.9.1=pypi_0
protobuf=3.20.3=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycm=4.0=pypi_0
pycparser=2.21=pypi_0
pydyf=0.7.0=pypi_0
pygame=2.5.1=pypi_0
pygments=2.16.1=pypi_0
pyparsing=3.1.0=pypi_0
pyphen=0.14.0=pypi_0
python=3.7.16=h6244533_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3=pypi_0
pywavelets=1.3.0=pypi_0
pyyaml=5.3=pypi_0
pyyaml-env-tag=0.1=pypi_0
reportlab=4.0.4=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
rst2pdf=0.99=pypi_0
scikit-image=0.16.2=pypi_0
scikit-learn=0.23.1=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.10.0=pypi_0
setuptools=65.6.3=py37haa95532_0
six=1.15.0=pypi_0
smartypants=2.0.1=pypi_0
snowballstemmer=2.2.0=pypi_0
soupsieve=2.4.1=pypi_0
sphinx=5.3.0=pypi_0
sphinx-basic-ng=1.0.0b2=pypi_0
sphinx-rtd-theme=1.3.0=pypi_0
sphinx-simplepdf=1.6.0=pypi_0
sphinxcontrib-applehelp=1.0.2=pypi_0
sphinxcontrib-devhelp=1.0.2=pypi_0
sphinxcontrib-htmlhelp=2.0.0=pypi_0
sphinxcontrib-jquery=4.1=pypi_0
sphinxcontrib-jsmath=1.0.1=pypi_0
sphinxcontrib-napoleon=0.7=pypi_0
sphinxcontrib-qthelp=1.0.3=pypi_0
sphinxcontrib-serializinghtml=1.1.5=pypi_0
sqlite=3.41.2=h2bbff1b_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.5.0=pypi_0
tensorflow-estimator=2.5.0=pypi_0
termcolor=1.1.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.2.1=pypi_0
tqdm=4.43.0=pypi_0
typing-extensions=4.7.1=pypi_0
urllib3=1.26.16=pypi_0
vc=14.2=h21ff451_1
visualkeras=0.0.2=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=3.0.0=pypi_0
weasyprint=59.0=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py37haa95532_0
wincertstore=0.2=py37haa95532_2
wrapt=1.12.1=pypi_0
zipp=3.15.0=pypi_0
zopfli=0.2.2=pypi_0
